Added Exercise Day 3 solution task 1
